Changed filenames and added some additional currency and regional support
